FIXME policycoreutils: semanage: FIXME make add operations into modify
Scripts which use semanage often don't know if what they intend to do
has already been done.  Should they use a -a because it doesn't exist or
should they use -m because it does?  This patch just makes -a use -m if
an entry already exists.

I'm not certain this is good practice.  Why can't scripts tell the
reason and handle the error themselves rather than not have a choice in
the tools?  What if two different rpms wanted to install a new selinux
user and the creator of those rpms didn't know about the other one?
They'd have no way of knowing about the name collision and the fact that
policy was likely not what they intended....

NOT-Signed-off-by: Eric Paris <eparis@redhat.com>