FIXME libselinux: use running kernel policy by default FIXME
SELinux semanage tools that query SELinux policy should be using the
policy that is loaded into the kernel, rather then the policy on disk.
Since they may be different.  The -p option should be available to most
utilities to look at other policy stores.  This should transperently
fallback to the policy on disk if the in kernel policy is unavailable.

[audit2why does not fall back?  depends on binary policy path shit]

Signed-off-by: Eric Paris <eparis@redhat.com>
Acked-by: Dan Walsh <dwalsh@redhat.com>