libselinux: get_context_list: return failure when unable to order list
When a program, like sshd, calls get_ordered_context_list() they want
to know what domain to transition into.  We determine this by looking up
all of the domains they can transition into and then ordering that list
by:

/etc/selinux/targeted/contexts/users/dwalsh
and then
/etc/selinux/targeted/contexts/default_contexts

Previously if we could not find an ordered entry, we would just randomly
pick some entry off of the reachable list  and give it a whirl.  That
might work in permissive, but it's just a bad idea in enforcing.

Instead, if we cannot find a valid order, just use the failsafe context.
It should at least be predicatable and we won't end up with completely
random crap (I've seem things like all of userspace running as
modprobe_t) happening.  Randomness is not our friend.
Return an error in enforcing.

NOT-Signed-off-by: Eric Paris <eparis@redhat.com>